+++ deploy.sh
@@ -12,7 +12,7 @@ echo "Compressing Image..."
 docker save $IMAGE_NAME:latest | gzip > looper-deploy.tar.gz
 
 echo "Uploading Assets..."
-scp looper-deploy.tar.gz .auth-key $VPS_USER@$VPS_HOST:~
+scp looper-deploy.tar.gz .auth-key backup.sh $VPS_USER@$VPS_HOST:~
 
 echo "Remote: Loading & Restarting..."
 ssh $VPS_USER@$VPS_HOST << EOF